Fruity - a better optimization tool
Yup it's fresh, but that's the point. I think this is too important to skip over as it's simply a better tool than existing options in the list. This list is also how people find such tools, so the policy of not including fresh tools has the effect of making sure new tools are not discovered.

The reasoning for the development of this tool is given here: https://github.com/marcandre/fruity#approach-